fix film list block button at Dodaj film

+++ src/main/java/cinema/system/database/models/Films.java
@@ -88,7 +88,7 @@ public class Films implements BaseModel{
         this.isbn = isbn;
     }
 
-    public float getRating() {
+    public int getRating() {
         return rating;
     }
